15/08/2023

+++ src/axios/index.js
@@ -10,6 +10,11 @@ const aFetch = axios.create({
     timeout: 2000,
 });
 
+const urlWithAuth = [
+    '/user/auth/tg/',
+    '/user/auth/refresh/'
+]
+
 let access_token
 let refresh_token
 
@@ -35,23 +40,31 @@ async function saveTokensToDB(userAuth) {
 }
 
 
+aFetch.refresh = false
+
 aFetch.interceptors.request.use(async (c) => {
     await getTokensFromDB()
     c.headers.Authorization = access_token ? `Bearer ${access_token}` : ''
-    if (c.url.includes('/user/auth/remove/')) {
+    if (c.url.includes('/user/auth/remove/') && c.data.refresh_token) {
         await Promise.all([
             storeDB.removeElement(constants.store.STORE, ACCESS_TOKEN),
             storeDB.removeElement(constants.store.STORE, REFRESH_TOKEN)
         ])
     }
-    return c;
+    if (aFetch.refresh) {
+        return new Promise((res) => {
+            setTimeout(() => res(c), 1500)
+        })
+    } else {
+        return c;
+    }
 }, err => console.error(err))
 
 
 aFetch.interceptors.response.use(
     response => {
         const url = response.config.url
-        if (url.includes('/user/auth/')) {
+        if (urlWithAuth.includes(url)) {
             const {ok, data} = response.data
             if (ok) {
                 saveTokensToDB(data)
@@ -64,7 +77,7 @@ aFetch.interceptors.response.use(
         const originalRequest = err.config;
         if (err.response.status === 401 && !originalRequest._retry) {
             originalRequest._retry = true;
-
+            aFetch.refresh = true
             return axios.get(baseURL + '/user/auth/refresh/', {
                 headers: {
                     Authorization: refresh_token ? `Bearer ${refresh_token}` : '',
@@ -80,15 +93,15 @@ aFetch.interceptors.response.use(
                         saveTokensToDB(userAuth)
                             .catch(console.error)
                         originalRequest.headers['Authorization'] = `Bearer ${userAuth.token}`;
-                        return axios(originalRequest);
                     })
+                    return axios(originalRequest);
                 } else if (window) {
                     window.localStorage.setItem(USER_AUTH, JSON.stringify(null))
                 } else if (postMessage) {
                     postMessage({type: UNAUTHORIZED})
                 }
                 return originalRequest
-            })
+            }).finally(() => aFetch.refresh = false)
         }
         return Promise.reject(err);
     })
